laptop_be/requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.4.3
openpyxl==3.1.5
pandas==3.0.1
psycopg==3.3.3
psycopg-binary==3.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.48
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
